Functions to normalize, transform, measure distance between numeric vectors

transform_cosine is the cosine transformation. transform_logistic is the logistic transformation. transform_zscore is the zscore transformation. dist_canberra computes the Canberra distance between 2 numeric vectors. dist_cosine computes the cosine angle distance between 2 numeric vectors. dist_euclidean compute the Euclidience distance between 2 numeric vectors. dist_pearson compute the Pearson correlation distance between 2 numeric vectors.

if(!is.vector(x)) abort("x must be vector")

transform_cosine(x, max = 100)

transform_logistic(x, max = 100)

transform_zscore(x)

dist_canberra(x, y)

dist_cosine(x, y)

dist_euclidean(x, y)

dist_pearson(x, y)

trim_outlier(x, fraction = 0.01)

normalize_percentile(x, fraction = 0.01)

get_confidence_interval(x, level = 0.95)

decile_band(x, band_ptile = c(seq(0, 0.95, 0.05)))

decile_ptile(x, band_ptile = c(seq(0, 0.95, 0.05)))

rank_ptile(x, level_rank = c(1, 2, 3, 4, seq(5, 100, 5)))

dc_mode(x, na.rm = FALSE)

dc_ceiling(x, digits = 0, na.rm = FALSE)

Arguments

x
A numeric vector

max
A numeric value

y
A numeric vector

fraction
The percentile value (0 to 0.5) to trim out

level
The CI level (0.5 to 1.0) of observations to be measured.

band_ptile
The percentail band (0.0 to 1.0)

level_rank
The rank level (0.0 to 1.0) for calculating percentile

na.rm
A logical value indicating whether NA values should be stripped before the computation proceeds.

digits
similar to rbase::round() which is integer indicating the number of decimal places (round) or significant digits (signif) to be used. Negative values are allowed

Value

returns a numeric vector after normaliztion or distance between 2 vectors.

Details

decile_band add columns with decile bands decile_ptile add columns with decile percentiles rank_ptile add columns with ranked percentiles

dc_ceiling similar to rbase::ceiling() with support decimal round up dc_mode compute the stats mode

Examples

library(dacol)
library(dplyr)

#> 
#> Attaching package: 'dplyr'

#> The following object is masked from 'package:testthat':
#> 
#>     matches

#> The following objects are masked from 'package:stats':
#> 
#>     filter, lag

#> The following objects are masked from 'package:base':
#> 
#>     intersect, setdiff, setequal, union

max = 30
dta1 = tibble(x1 = seq(-1.2*max, 1.2*max, length.out = 200),
              x2 = seq(1, max, length.out = 200),
              x3 = sample(200))

dta1 = mutate(dta1,

              # Transformation
              y_cosine   = transform_cosine(x1, max),
              y_logistic = transform_logistic(x2, max),
              y_zcore    = transform_zscore(x2),

              # Distant between 2 vector columns
              y_dist_canb = dist_canberra(x2, x3),
              y_dist_cos  = dist_cosine(x2, y_zcore),
              y_dist_euc  = dist_euclidean(x2, y_zcore),
              y_dist_pear = dist_pearson(x2, y_zcore),

              # Manage outliers
              y_trim = trim_outlier(x3, 0.01),
              y_norm = normalize_percentile(x3, 0.01),

              # Stats measures
              y_mode = dc_mode(x3),
              y_ceil = dc_ceiling(x3, -1),

              # Band segmentation
              y_dec_band1 = decile_band(x3),
              y_dec_band2 = decile_band(x3, c(seq(0, 0.9, 0.1))),
              y_dec_ptile1 = decile_ptile(x3),
              y_dec_ptile2 = decile_ptile(x3, c(seq(0, 0.9, 0.1)))
              )